# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CRUDL API for subscription records.",
        "title": "Subscriptions API",
        "contact": {}
    },
    "basePath": "/v1",
    "paths": {
        "/health": {
            "get": {
                "description": "Returns service and database health status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.HealthSuccessResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/controllers.HealthErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscription-cost": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Sums monthly prices for subscriptions active during the period. Optional filters: user_id, service_name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription-cost"
                ],
                "summary": "Total subscription cost for a period",
                "parameters": [
                    {
                        "type": "string",
                        "default": "01-2024",
                        "description": "Period start (MM-YYYY)",
                        "name": "period_start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "12-2024",
                        "description": "Period end (MM-YYYY)",
                        "name": "period_end",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "60601fee-2bf1-4721-ae6f-7636e79a0cba",
                        "description": "Filter by user UUID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "Yandex Plus",
                        "description": "Filter by subscription service name",
                        "name": "service_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.SubscriptionCostSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIUnauthorizedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIInternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscriptions/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a subscription record.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Create subscription",
                "parameters": [
                    {
                        "description": "Subscription payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SubscriptionToSave"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.SubscriptionSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIUnauthorizedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIInternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscriptions/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Delete subscription",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Subscription ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.SubscriptionDeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIUnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.APINotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIInternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscriptions/get/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Get subscription",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Subscription ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.SubscriptionSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIUnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.APINotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIInternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscriptions/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a paginated list. Optional filter by user_id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "List subscriptions",
                "parameters": [
                    {
                        "type": "string",
                        "default": "60601fee-2bf1-4721-ae6f-7636e79a0cba",
                        "description": "Filter by user UUID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIUnauthorizedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIInternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscriptions/update/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Update subscription",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Subscription ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SubscriptionToUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.SubscriptionSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIUnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.APINotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIInternalErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.APIBadRequestResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/controllers.apiErrorData"
                        }
                    ],
                    "x-nullable": true
                },
                "message": {
                    "type": "string",
                    "example": "invalid subscription id"
                },
                "status": {
                    "type": "string",
                    "example": "failed"
                }
            }
        },
        "controllers.APIInternalErrorResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/controllers.apiErrorData"
                        }
                    ],
                    "x-nullable": true
                },
                "message": {
                    "type": "string",
                    "example": "database error"
                },
                "status": {
                    "type": "string",
                    "example": "failed"
                }
            }
        },
        "controllers.APINotFoundResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/controllers.apiErrorData"
                        }
                    ],
                    "x-nullable": true
                },
                "message": {
                    "type": "string",
                    "example": "subscription not found"
                },
                "status": {
                    "type": "string",
                    "example": "failed"
                }
            }
        },
        "controllers.APIUnauthorizedResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/controllers.apiErrorData"
                        }
                    ],
                    "x-nullable": true
                },
                "message": {
                    "type": "string",
                    "example": "no or empty x-api-key"
                },
                "status": {
                    "type": "string",
                    "example": "failed"
                }
            }
        },
        "controllers.HealthErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "Database ping failed"
                },
                "status": {
                    "type": "string",
                    "example": "unhealthy"
                }
            }
        },
        "controllers.HealthSuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Service is running"
                },
                "status": {
                    "type": "string",
                    "example": "healthy"
                },
                "version": {
                    "type": "string",
                    "example": "dev"
                }
            }
        },
        "controllers.SubscriptionCostSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.SubscriptionCostResult"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "controllers.SubscriptionDeleteResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "subscription deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "controllers.SubscriptionSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Subscription"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "controllers.apiErrorData": {
            "type": "object"
        },
        "models.Subscription": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "12-2025"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "price": {
                    "type": "integer",
                    "example": 400
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "start_date": {
                    "type": "string",
                    "example": "01-2024"
                },
                "user_id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                }
            }
        },
        "models.SubscriptionCostResult": {
            "type": "object",
            "properties": {
                "period_end": {
                    "type": "string",
                    "example": "12-2024"
                },
                "period_start": {
                    "type": "string",
                    "example": "01-2024"
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "total": {
                    "type": "integer",
                    "example": 4800
                },
                "user_id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                }
            }
        },
        "models.SubscriptionToSave": {
            "type": "object",
            "required": [
                "service_name",
                "start_date",
                "user_id"
            ],
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "12-2025"
                },
                "price": {
                    "type": "number",
                    "minimum": 0,
                    "example": 199.75
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "start_date": {
                    "type": "string",
                    "example": "01-2024"
                },
                "user_id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                }
            }
        },
        "models.SubscriptionToUpdate": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "12-2025"
                },
                "price": {
                    "type": "number",
                    "minimum": 0,
                    "example": 599
                },
                "service_name": {
                    "type": "string",
                    "example": "Netflix"
                },
                "start_date": {
                    "type": "string",
                    "example": "03-2024"
                },
                "user_id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        }
    }
}